+++ ksa_compliance/hooks.py
@@ -13,6 +13,7 @@ app_license = 'Copyright (c) 2023 LavaLoon'
 # app_include_css = "/assets/ksa_compliance/css/ksa_compliance.css"
 app_include_js = [
     '/assets/ksa_compliance/js/feedback_dialog.js',
+    '/assets/ksa_compliance/js/premium_announcement.js'
 ]
 # include js, css files in header of web template
 # web_include_css = "/assets/ksa_compliance/css/ksa_compliance.css"
@@ -30,7 +31,7 @@ app_include_js = [
 
 # include js in doctype views
 # doctype_js = {"doctype" : "public/js/doctype.js"}
-# doctype_list_js = {"doctype" : "public/js/doctype_list.js"}
+doctype_list_js = {"Sales Invoice" : "public/js/sales_invoice_list.js"}
 # doctype_tree_js = {"doctype" : "public/js/doctype_tree.js"}
 # doctype_calendar_js = {"doctype" : "public/js/doctype_calendar.js"}