+++ ksa_compliance/ksa_compliance/doctype/sales_invoice_additional_fields/sales_invoice_additional_fields_list.js
@@ -1,0 +1,5 @@
+frappe.listview_settings["Sales Invoice Additional Fields"] = {
+    onload(listview) {
+        ksa_compliance.premium_announcement.show_announcement(listview)
+    }
+}