+++ ksa_compliance/public/js/sales_invoice_list.js
@@ -1,0 +1,5 @@
+frappe.listview_settings['Sales Invoice'] = {
+    onload(listview) {
+        ksa_compliance.premium_announcement.show_announcement(listview)
+    }
+}